Update HigestNo.java
I have added a separate else block if in case the user enters equal numbers. That will let the user know to enter different numbers. Also, I have used else if, which makes the code more efficient. The problem with if statements is that the compiler will go to the next if statement even though the earlier if the statement was satisfied. When we use else if, the compiler comes of the conditional statements when it finds that the statement which fulfills the condition. In your case, if num1 is greater, then the compiler will go to 2nd and 3rd if statement also because you have used if statements. When we use else if, if num1 is greater, then the compiler won't check the remaining conditions. The use of if statement and else if statements will give the same output, however, else if is more efficient than if.